+++ Tests/PeripheryTests/RetentionTest.swift
@@ -347,7 +347,7 @@ final class RetentionTest: SourceGraphTestCase {
         }
     }
 
-    func testUnusedOverridenMethod() {
+    func testUnusedOverriddenMethod() {
         analyze(retainPublic: true) {
             assertReferenced(.class("FixtureClass81Base")) {
                 self.assertNotReferenced(.functionMethodInstance("someMethod()"))
@@ -358,7 +358,7 @@ final class RetentionTest: SourceGraphTestCase {
         }
     }
 
-    func testOverridenMethodRetainedBySuper() {
+    func testOverriddenMethodRetainedBySuper() {
         analyze(retainPublic: true) {
             assertReferenced(.class("FixtureClass82Base")) {
                 self.assertReferenced(.functionMethodInstance("someMethod()"))
@@ -881,6 +881,17 @@ final class RetentionTest: SourceGraphTestCase {
                 self.assertReferenced(.varInstance("assignOnlyProperty"))
                 self.assertNotAssignOnlyProperty(.varInstance("assignOnlyProperty"))
             }
+            assertReferenced(.class("FixtureClass212")) {
+                self.assertReferenced(.functionMethodInstance("protocolFunc(param:)")) {
+                    self.assertReferenced(.varParameter("param"))
+                }
+            }
+            assertReferenced(.class("FixtureClass213")) {
+                self.assertReferenced(.functionMethodInstance("someFunc(a:b:c:)")) {
+                    self.assertReferenced(.varParameter("b"))
+                    self.assertReferenced(.varParameter("c"))
+                }
+            }
             assertReferenced(.class("Fixture205"))
             assertReferenced(.protocol("Fixture205Protocol"))
             assertNotRedundantProtocol("Fixture205Protocol")
@@ -1124,13 +1135,26 @@ final class RetentionTest: SourceGraphTestCase {
                 }
 
                 // Used in override.
-                self.assertReferenced(.functionMethodInstance("func4(param:)")) {
+                // func4(param: String)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 14)) {
                     self.assertReferenced(.varParameter("param"))
+                }
+
+                // Not used in any function.
+                // func4(param: Int)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 17)) {
+                    self.assertNotReferenced(.varParameter("param"))
                 }
 
                 // Not used in any function.
                 self.assertReferenced(.functionMethodInstance("func5(param:)")) {
                     self.assertNotReferenced(.varParameter("param"))
+                }
+
+                // Overridden in multiple subclass branches.
+                self.assertReferenced(.functionMethodInstance("func7(param1:param2:)")) {
+                    self.assertReferenced(.varParameter("param1"))
+                    self.assertNotReferenced(.varParameter("param2"))
                 }
             }
 
@@ -1144,17 +1168,50 @@ final class RetentionTest: SourceGraphTestCase {
                 self.assertReferenced(.functionMethodInstance("func3(param:)")) {
                     self.assertUsedParameter("param")
                 }
+
+                // Not used in any function.
+                // func4(param: Int)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 36)) {
+                    self.assertNotReferenced(.varParameter("param"))
+                }
+
+                // Overridden in multiple subclass branches.
+                self.assertReferenced(.functionMethodInstance("func7(param1:param2:)")) {
+                    self.assertUsedParameter("param1")
+                    self.assertNotReferenced(.varParameter("param2"))
+                }
             }
 
             assertReferenced(.class("FixtureClass101Subclass2")) {
                 // The param is used.
-                self.assertReferenced(.functionMethodInstance("func4(param:)")) {
+                // func4(param: String)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 44)) {
                     self.assertUsedParameter("param")
+                }
+
+                // Not used in any function.
+                // func4(param: Int)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 48)) {
+                    self.assertNotReferenced(.varParameter("param"))
                 }
 
                 // Not used in any function.
                 self.assertReferenced(.functionMethodInstance("func5(param:)")) {
                     self.assertNotReferenced(.varParameter("param"))
+                }
+
+                // Overridden in multiple subclass branches.
+                self.assertReferenced(.functionMethodInstance("func7(param1:param2:)")) {
+                    self.assertReferenced(.varParameter("param1"))
+                    self.assertNotReferenced(.varParameter("param2"))
+                }
+            }
+
+            assertReferenced(.class("FixtureClass101Subclass3")) {
+                // Overridden in multiple subclass branches.
+                self.assertReferenced(.functionMethodInstance("func7(param1:param2:)")) {
+                    self.assertReferenced(.varParameter("param1"))
+                    self.assertNotReferenced(.varParameter("param2"))
                 }
             }
 
@@ -1255,8 +1312,15 @@ final class RetentionTest: SourceGraphTestCase {
                 }
 
                 // Unused in extension, but used in conformance.
-                self.assertReferenced(.functionMethodInstance("func4(param:)")) {
+                // func4(param: String)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 11)) {
                     self.assertReferenced(.varParameter("param"))
+                }
+
+                // Unused.
+                // func4(param: Int)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 13)) {
+                    self.assertNotReferenced(.varParameter("param"))
                 }
 
                 // Used in a conformance.
@@ -1277,8 +1341,15 @@ final class RetentionTest: SourceGraphTestCase {
                 }
 
                 // Used in a conformance by another class.
-                self.assertReferenced(.functionMethodInstance("func4(param:)")) {
+                // func4(param: String)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 27)) {
                     self.assertReferenced(.varParameter("param"))
+                }
+
+                // Unused.
+                // func4(param: Int)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 28)) {
+                    self.assertNotReferenced(.varParameter("param"))
                 }
             }
 
@@ -1331,8 +1402,15 @@ final class RetentionTest: SourceGraphTestCase {
                 }
 
                 // The param is used.
-                self.assertReferenced(.functionMethodInstance("func4(param:)")) {
+                // func4(param: String)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 50)) {
                     self.assertUsedParameter("param")
+                }
+
+                // Unused.
+                // func4(param: Int)
+                self.assertReferenced(.functionMethodInstance("func4(param:)", line: 54)) {
+                    self.assertNotReferenced(.varParameter("param"))
                 }
 
                 // The param is used.